GH_GQL renaoming conventions
