+++ src/server/services/gql/reposWithCommitCountsService.ts
@@ -1,5 +1,5 @@
-import { GH_GQL_getReposWithCommitCounts } from '@/server/lib/gh-gql/reposWithCommitCounts'
-import { GH_GQL_getViewer } from '@/server/lib/gh-gql/users'
+import { GH_GQL_getReposWithCommitCounts } from '@/server/lib/gh-gql/AllRepoCommitCounts'
+import { GH_GQL_Call__Viewer } from '@/server/lib/gh-gql/Viewer'
 import { SHARED_APIFields_GetLifetimeStats } from '@/shared/models'
 
 // const processReposWithCommits = (repo: GH_GQL_RepoConnection): SHARED_Model_LifetimeStats => {
@@ -8,13 +8,13 @@ import { SHARED_APIFields_GetLifetimeStats } from '@/shared/models'
 // }
 
 export const getReposWithCommitCounts = async (accessToken: string): Promise<SHARED_APIFields_GetLifetimeStats> => {
-    const { success: userSuccess, error: userError, user } = await GH_GQL_getViewer(accessToken)
+    const { success: viewerSuccess, error: viewerError, viewer } = await GH_GQL_Call__Viewer(accessToken)
 
-    if (!userSuccess || user === undefined) {
-        return { stats: undefined, success: false, error: userError }
+    if (!viewerSuccess || viewer === undefined) {
+        return { stats: undefined, success: false, error: viewerError }
     }
 
-    const { id } = user
+    const { id } = viewer
     const {
         success: repoSuccess,
         error: repoError,
